+++ src/main/java/com/lampirg/calculator/logic/expression/Divide.java
@@ -3,7 +3,7 @@ package com.lampirg.calculator.logic.expression;
 import java.util.function.BiFunction;
 
 public class Divide extends BinaryExpression {
-    public Divide(int x1, int x2) {
+    public Divide(double x1, double x2) {
         super(x1, x2, (n1, n2) -> n1 / n2);
     }
 }
